refactor(stream): Be extra clear that macros mod is not for use
