+++ crates/anstream/src/_macros.rs
